api/retriever/requirements.txt
fastapi
uvicorn
sentence-transformers
pymilvus
python-dotenv
tqdm
